move mockclock to testhelpers
